+++ skrub/__init__.py
@@ -22,7 +22,6 @@ from ._gap_encoder import GapEncoder
 from ._minhash_encoder import MinHashEncoder
 from ._similarity_encoder import SimilarityEncoder
 from ._table_vectorizer import SuperVectorizer, TableVectorizer
-from ._target_encoder import TargetEncoder
 
 with open(_Path(__file__).parent / "VERSION.txt") as _fh:
     __version__ = _fh.read().strip()
@@ -37,7 +36,28 @@ __all__ = [
     "SimilarityEncoder",
     "SuperVectorizer",
     "TableVectorizer",
-    "TargetEncoder",
     "deduplicate",
     "compute_ngram_distance",
 ]
+
+
+def __getattr__(name):
+    if name == "TargetEncoder":
+        import warnings
+        import sklearn
+        from ._utils import parse_version
+
+        if parse_version(sklearn.__version__) >= parse_version("1.4.0"):
+            warnings.warn(
+                "Importing TargetEncoder from skrub is deprecated and will be "
+                "removed in a future version. It should be imported from sklearn."
+            )
+            from sklearn.preprocessing import TargetEncoder
+        else:
+            from ._target_encoder import TargetEncoder
+        return TargetEncoder
+    try:
+        return globals()[name]
+    except KeyError:
+        # This is turned into the appropriate ImportError
+        raise AttributeError