+++ skrub/_target_encoder.py
@@ -1,322 +1,402 @@
-import collections
-from typing import Literal
+# Vendored implementation of the TargetEncoder from scikit-learn.
+# TODO remove it when sklearn min version >= 1.4.0.
 
 import numpy as np
-from numpy.typing import ArrayLike, NDArray
-from sklearn.base import BaseEstimator, TransformerMixin
+
+from sklearn.base import OneToOneFeatureMixin
+from sklearn.utils.multiclass import type_of_target
+from sklearn.utils.validation import _check_y, check_consistent_length
+from sklearn.preprocessing._encoders import _BaseEncoder
 from sklearn.preprocessing import LabelEncoder
-from sklearn.utils import check_array
-from sklearn.utils.fixes import _object_dtype_isnan
-from sklearn.utils.validation import check_is_fitted
-
-from skrub._utils import check_input
+from sklearn.model_selection import KFold, StratifiedKFold
 
 
-def lambda_(x, n):
-    return x / (x + n)
+def _fit_encoding_fast(X_int, y, n_categories, smooth, y_mean):
+    """Fit a target encoding on X_int and y.
+
+    This implementation uses Eq 7 from [1] to compute the encoding.
+    As stated in the paper, Eq 7 is the same as Eq 3.
+
+    [1]: Micci-Barreca, Daniele. "A preprocessing scheme for high-cardinality
+         categorical attributes in classification and prediction problems"
+    """
+    n_features = X_int.shape[1]
+    smooth_sum = smooth * y_mean
+
+    encodings = [
+        np.full(n_categories[feat_idx], fill_value=y_mean)
+        for feat_idx in range(n_features)
+    ]
+
+    for feat_idx in range(n_features):
+        n_cats = n_categories[feat_idx]
+        col = X_int[:, feat_idx]
+        col_known = col[col != -1]
+
+        sums = np.bincount(col_known, weights=y, minlength=n_cats) + smooth_sum
+        counts = np.bincount(col_known, minlength=n_cats) + smooth
+
+        np.divide(sums, counts, out=encodings[feat_idx], where=counts != 0)
+
+    return encodings
 
 
-class TargetEncoder(BaseEstimator, TransformerMixin):
-    """Encode categorical features as a numeric array given a target vector.
+def _fit_encoding_fast_auto_smooth(X_int, y, n_categories, y_mean, y_variance):
+    """Fit a target encoding on X_int and y with auto smoothing.
 
-    Each category is encoded given the effect that it has in the
-    target variable :term:`y`. The method considers that categorical
-    variables can present rare categories. It represents each category by the
-    probability of :term:`y` conditional on this category.
-    In addition, it takes an empirical Bayes approach to shrink the estimate.
+    This implementation uses Eq 5 and 6 from [1].
+
+    [1]: Micci-Barreca, Daniele. "A preprocessing scheme for high-cardinality
+         categorical attributes in classification and prediction problems"
+    """
+    n_features = X_int.shape[1]
+
+    encodings = [np.empty(n_categories[feat_idx]) for feat_idx in range(n_features)]
+
+    for feat_idx in range(n_features):
+        n_cats = n_categories[feat_idx]
+
+        col = X_int[:, feat_idx]
+        col_known = col[col != -1]
+
+        counts = np.bincount(col_known, minlength=n_cats)
+        means = np.bincount(col_known, weights=y, minlength=n_cats)
+        np.divide(means, counts, out=means, where=counts != 0)
+
+        diff = y - means[col]
+        sum_of_squared_diffs = np.bincount(col, weights=diff * diff, minlength=n_cats)
+        np.divide(
+            sum_of_squared_diffs, counts, out=sum_of_squared_diffs, where=counts != 0
+        )
+
+        lambda_ = y_variance * counts
+        np.divide(
+            lambda_, lambda_ + sum_of_squared_diffs, out=lambda_, where=lambda_ != 0
+        )
+
+        encodings[feat_idx][:] = lambda_ * means + (1 - lambda_) * y_mean
+
+    return encodings
+
+
+class TargetEncoder(OneToOneFeatureMixin, _BaseEncoder):
+    """Target Encoder for regression and classification targets.
+
+    Each category is encoded based on a shrunk estimate of the average target
+    values for observations belonging to the category. The encoding scheme mixes
+    the global target mean with the target mean conditioned on the value of the
+    category. [MIC]_
+
+    :class:`TargetEncoder` considers missing values, such as `np.nan` or `None`,
+    as another category and encodes them like any other category. Categories
+    that are not seen during :meth:`fit` are encoded with the target mean, i.e.
+    `target_mean_`.
+
+    Read more in the :ref:`User Guide <target_encoder>`.
+
+    .. note::
+        `fit(X, y).transform(X)` does not equal `fit_transform(X, y)` because a
+        cross fitting scheme is used in `fit_transform` for encoding. See the
+        :ref:`User Guide <target_encoder>` for details.
+
+    .. versionadded:: 1.3
 
     Parameters
     ----------
-    categories : 'auto' or list of list of int or str
+    categories : "auto" or list of shape (n_features,) of array-like, default="auto"
         Categories (unique values) per feature:
 
-        - 'auto' : Determine categories automatically from the training data.
-        - list : `categories[i]` holds the categories expected in the `i`-th
-          column. The passed categories must be sorted and should not mix
-          strings and numeric values.
+        - `"auto"` : Determine categories automatically from the training data.
+        - list : `categories[i]` holds the categories expected in the i-th column. The
+          passed categories should not mix strings and numeric values within a single
+          feature, and should be sorted in case of numeric values.
 
-        The categories used can be found in the ``categories_`` attribute.
-    clf_type : {'regression', 'binary-clf', 'multiclass-clf'}, default='binary-clf'
-        The type of classification/regression problem.
-    dtype : number type, default=np.float64
-        Desired dtype of output.
-    handle_unknown : {'error', 'ignore'}, default='error'
-        Whether to raise an error or ignore if an unknown categorical feature is
-        present during transform (default is to raise). When this parameter
-        is set to 'ignore' and an unknown category is encountered during
-        transform, the encoded columns for this feature
-        will be assigned the prior mean of the target variable.
-    handle_missing : {'error', ''}, default=''
-        Whether to raise an error or impute with blank string '' if missing
-        values (NaN) are present during TargetEncoder.fit
-        (default is to impute).
-        When this parameter is set to '', and a missing value is encountered
-        during TargetEncoder.fit_transform, the resulting encoded
-        columns for this feature will be all zeros.
+        The used categories are stored in the `categories_` fitted attribute.
+
+    target_type : {"auto", "continuous", "binary"}, default="auto"
+        Type of target.
+
+        - `"auto"` : Type of target is inferred with
+          :func:`~sklearn.utils.multiclass.type_of_target`.
+        - `"continuous"` : Continuous target
+        - `"binary"` : Binary target
+
+        .. note::
+            The type of target inferred with `"auto"` may not be the desired target
+            type used for modeling. For example, if the target consisted of integers
+            between 0 and 100, then :func:`~sklearn.utils.multiclass.type_of_target`
+            will infer the target as `"multiclass"`. In this case, setting
+            `target_type="continuous"` will specify the target as a regression
+            problem. The `target_type_` attribute gives the target type used by the
+            encoder.
+
+    smooth : "auto" or float, default="auto"
+        The amount of mixing of the target mean conditioned on the value of the
+        category with the global target mean. A larger `smooth` value will put
+        more weight on the global target mean.
+        If `"auto"`, then `smooth` is set to an empirical Bayes estimate.
+
+    cv : int, default=5
+        Determines the number of folds in the cross fitting strategy used in
+        :meth:`fit_transform`. For classification targets, `StratifiedKFold` is used
+        and for continuous targets, `KFold` is used.
+
+    shuffle : bool, default=True
+        Whether to shuffle the data in :meth:`fit_transform` before splitting into
+        folds. Note that the samples within each split will not be shuffled.
+
+    random_state : int, RandomState instance or None, default=None
+        When `shuffle` is True, `random_state` affects the ordering of the
+        indices, which controls the randomness of each fold. Otherwise, this
+        parameter has no effect.
+        Pass an int for reproducible output across multiple function calls.
+        See :term:`Glossary <random_state>`.
 
     Attributes
     ----------
+    encodings_ : list of shape (n_features,) of ndarray
+        Encodings learnt on all of `X`.
+        For feature `i`, `encodings_[i]` are the encodings matching the
+        categories listed in `categories_[i]`.
+
+    categories_ : list of shape (n_features,) of ndarray
+        The categories of each feature determined during fitting or specified
+        in `categories`
+        (in order of the features in `X` and corresponding with the output
+        of :meth:`transform`).
+
+    target_type_ : str
+        Type of target.
+
+    target_mean_ : float
+        The overall mean of the target. This value is only used in :meth:`transform`
+        to encode categories.
+
     n_features_in_ : int
-        Number of features in the data seen during TargetEncoder.fit.
-    categories_ : list of ndarray
-        The categories of each feature determined during TargetEncoder.fit
-        (in order corresponding with output of TargetEncoder.transform).
-    n_ : int
-        Length of :term:`y`
+        Number of features seen during :term:`fit`.
+
+    feature_names_in_ : ndarray of shape (`n_features_in_`,)
+        Names of features seen during :term:`fit`. Defined only when `X`
+        has feature names that are all strings.
 
     See Also
     --------
-    GapEncoder
-        Encodes dirty categories (strings) by constructing latent topics with
-        continuous encoding.
-    MinHashEncoder
-        Encode string columns as a numeric array with the minhash method.
-    SimilarityEncoder
-        Encode string columns as a numeric array with n-gram string similarity.
+    OrdinalEncoder : Performs an ordinal (integer) encoding of the categorical features.
+        Contrary to TargetEncoder, this encoding is not supervised. Treating the
+        resulting encoding as a numerical features therefore lead arbitrarily
+        ordered values and therefore typically lead to lower predictive performance
+        when used as preprocessing for a classifier or regressor.
+    OneHotEncoder : Performs a one-hot encoding of categorical features. This
+        unsupervised encoding is better suited for low cardinality categorical
+        variables as it generate one new feature per unique category.
 
     References
     ----------
-    For more details, see Micci-Barreca, 2001: A preprocessing scheme for
-    high-cardinality categorical attributes in classification and prediction
-    problems.
+    .. [MIC] :doi:`Micci-Barreca, Daniele. "A preprocessing scheme for high-cardinality
+       categorical attributes in classification and prediction problems"
+       SIGKDD Explor. Newsl. 3, 1 (July 2001), 27–32. <10.1145/507533.507538>`
 
     Examples
     --------
-    >>> enc = TargetEncoder(handle_unknown='ignore')
-    >>> X = [['male'], ['Male'], ['Female'], ['male'], ['Female']]
-    >>> y = np.array([1, 2, 3, 4, 5])
+    With `smooth="auto"`, the smoothing parameter is set to an empirical Bayes estimate:
 
-    >>> enc.fit(X, y)
-    TargetEncoder(handle_unknown='ignore')
+    >>> import numpy as np
+    >>> from sklearn.preprocessing import TargetEncoder
+    >>> X = np.array([["dog"] * 20 + ["cat"] * 30 + ["snake"] * 38], dtype=object).T
+    >>> y = [90.3] * 5 + [80.1] * 15 + [20.4] * 5 + [20.1] * 25 + [21.2] * 8 + [49] * 30
+    >>> enc_auto = TargetEncoder(smooth="auto")
+    >>> X_trans = enc_auto.fit_transform(X, y)
 
-    The encoder has found the following categories:
+    >>> # A high `smooth` parameter puts more weight on global mean on the categorical
+    >>> # encodings:
+    >>> enc_high_smooth = TargetEncoder(smooth=5000.0).fit(X, y)
+    >>> enc_high_smooth.target_mean_
+    44...
+    >>> enc_high_smooth.encodings_
+    [array([44..., 44..., 44...])]
 
-    >>> enc.categories_
-    [array(['Female', 'Male', 'male'], dtype='<U6')]
-
-    We will encode the following categories, of which the first two are unknown :
-
-    >>> X2 = [['MALE'], ['FEMALE'], ['Female'], ['male'], ['Female']]
-
-    >>> enc.transform(X2)
-    array([[3.        ],
-        [3.        ],
-        [3.54545455],
-        [2.72727273],
-        [3.54545455]])
-
-    As expected, they were encoded according to their influence on y.
-    The unknown categories were assigned the mean of the target variable.
+    >>> # On the other hand, a low `smooth` parameter puts more weight on target
+    >>> # conditioned on the value of the categorical:
+    >>> enc_low_smooth = TargetEncoder(smooth=1.0).fit(X, y)
+    >>> enc_low_smooth.encodings_
+    [array([20..., 80..., 43...])]
     """
-
-    n_features_in_: int
-    _label_encoders_: list[LabelEncoder]
-    categories_: list[NDArray]
-    n_: int
 
     def __init__(
         self,
-        *,
-        categories: Literal["auto"] | list[list[str] | NDArray] = "auto",
-        clf_type: Literal["regression", "binary-clf", "multiclass-clf"] = "binary-clf",
-        dtype: type = np.float64,
-        handle_unknown: Literal["error", "ignore"] = "error",
-        handle_missing: Literal["error", ""] = "",
+        categories="auto",
+        target_type="auto",
+        smooth="auto",
+        cv=5,
+        shuffle=True,
+        random_state=None,
     ):
         self.categories = categories
-        self.dtype = dtype
-        self.clf_type = clf_type
-        self.handle_unknown = handle_unknown
-        self.handle_missing = handle_missing
+        self.smooth = smooth
+        self.target_type = target_type
+        self.cv = cv
+        self.shuffle = shuffle
+        self.random_state = random_state
 
-    def _more_tags(self) -> dict[str, list[str]]:
-        """
-        Used internally by sklearn to ease the estimator checks.
-        """
-        return {"X_types": ["categorical"]}
-
-    def fit(self, X: ArrayLike, y: ArrayLike) -> "TargetEncoder":
-        """Fit the instance to `X`.
+    def fit(self, X, y):
+        """Fit the :class:`TargetEncoder` to X and y.
 
         Parameters
         ----------
-        X : array-like, shape [n_samples, n_features]
+        X : array-like of shape (n_samples, n_features)
             The data to determine the categories of each feature.
-        y : ndarray
-            The associated target vector.
+
+        y : array-like of shape (n_samples,)
+            The target data used to encode the categories.
 
         Returns
         -------
-        TargetEncoder
-            Fitted TargetEncoder instance (self).
+        self : object
+            Fitted encoder.
         """
-        X = check_input(X)
-        self.n_features_in_ = X.shape[1]
-        if self.handle_missing not in ["error", ""]:
-            raise ValueError(
-                f"Got handle_missing={self.handle_missing!r}, but expected "
-                "any of {'error', ''}. "
-            )
-
-        mask = _object_dtype_isnan(X)
-        if mask.any():
-            if self.handle_missing == "error":
-                raise ValueError(
-                    "Found missing values in input data; set "
-                    "handle_missing='' to encode with missing values. "
-                )
-            else:
-                X[mask] = self.handle_missing
-
-        if self.handle_unknown not in ["error", "ignore"]:
-            raise ValueError(
-                f"Got handle_unknown={self.handle_unknown!r}, but expected "
-                "any of {'error', 'ignore'}. "
-            )
-
-        if self.categories != "auto":
-            for cats in self.categories:
-                if not np.all(np.sort(cats) == np.array(cats)):
-                    raise ValueError("Unsorted categories are not yet supported. ")
-
-        X_temp = check_array(X, dtype=None)
-        X = X_temp
-
-        n_samples, n_features = X.shape
-
-        self._label_encoders_ = [LabelEncoder() for _ in range(n_features)]
-
-        for j in range(n_features):
-            le = self._label_encoders_[j]
-            Xj = X[:, j]
-            if self.categories == "auto":
-                le.fit(Xj)
-            else:
-                if self.handle_unknown == "error":
-                    valid_mask = np.in1d(Xj, self.categories[j])
-                    if not np.all(valid_mask):
-                        diff = np.unique(Xj[~valid_mask])
-                        raise ValueError(
-                            f"Found unknown categories {diff} in column {j} during fit"
-                        )
-                le.classes_ = np.array(self.categories[j])
-
-        self.categories_ = [le.classes_ for le in self._label_encoders_]
-        self.n_ = len(y)
-        if self.clf_type in ["binary-clf", "regression"]:
-            self.Eyx_ = [
-                {cat: np.mean(y[X[:, j] == cat]) for cat in self.categories_[j]}
-                for j in range(len(self.categories_))
-            ]
-            self.Ey_ = np.mean(y)
-            self.counter_ = {j: collections.Counter(X[:, j]) for j in range(n_features)}
-        if self.clf_type in ["multiclass-clf"]:
-            self.classes_ = np.unique(y)
-
-            self.Eyx_ = {
-                c: [
-                    {
-                        cat: np.mean((y == c)[X[:, j] == cat])
-                        for cat in self.categories_[j]
-                    }
-                    for j in range(len(self.categories_))
-                ]
-                for c in self.classes_
-            }
-            self.Ey_ = {c: np.mean(y == c) for c in self.classes_}
-            self.counter_ = {j: collections.Counter(X[:, j]) for j in range(n_features)}
-        self.k_ = {j: len(self.counter_[j]) for j in self.counter_}
+        self._fit_encodings_all(X, y)
         return self
 
-    def transform(self, X: ArrayLike) -> NDArray:
-        """Transform `X` using the specified encoding scheme.
+    def fit_transform(self, X, y):
+        """Fit :class:`TargetEncoder` and transform X with the target encoding.
+
+        .. note::
+            `fit(X, y).transform(X)` does not equal `fit_transform(X, y)` because a
+            cross fitting scheme is used in `fit_transform` for encoding. See the
+            :ref:`User Guide <target_encoder>`. for details.
 
         Parameters
         ----------
-        X : array-like, shape [n_samples, n_features_new]
-            The data to encode.
+        X : array-like of shape (n_samples, n_features)
+            The data to determine the categories of each feature.
+
+        y : array-like of shape (n_samples,)
+            The target data used to encode the categories.
 
         Returns
         -------
-        2-d ndarray
+        X_trans : ndarray of shape (n_samples, n_features)
             Transformed input.
         """
-        check_is_fitted(self, attributes=["n_features_in_"])
-        X = check_input(X)
-        if X.shape[1] != self.n_features_in_:
-            raise ValueError(
-                f"The number of features in the input data ({X.shape[1]}) "
-                "does not match the number of features "
-                f"seen during fit ({self.n_features_in_})."
+        X_ordinal, X_known_mask, y, n_categories = self._fit_encodings_all(X, y)
+
+        # The cv splitter is voluntarily restricted to *KFold to enforce non
+        # overlapping validation folds, otherwise the fit_transform output will
+        # not be well-specified.
+        if self.target_type_ == "continuous":
+            cv = KFold(self.cv, shuffle=self.shuffle, random_state=self.random_state)
+        else:
+            cv = StratifiedKFold(
+                self.cv, shuffle=self.shuffle, random_state=self.random_state
             )
-        mask = _object_dtype_isnan(X)
-        if mask.any():
-            if self.handle_missing == "error":
-                raise ValueError(
-                    "Found missing values in input data; set "
-                    "handle_missing='' to encode with missing values. "
+
+        X_out = np.empty_like(X_ordinal, dtype=np.float64)
+        X_unknown_mask = ~X_known_mask
+
+        for train_idx, test_idx in cv.split(X, y):
+            X_train, y_train = X_ordinal[train_idx, :], y[train_idx]
+            y_mean = np.mean(y_train)
+
+            if self.smooth == "auto":
+                y_variance = np.var(y_train)
+                encodings = _fit_encoding_fast_auto_smooth(
+                    X_train, y_train, n_categories, y_mean, y_variance
                 )
             else:
-                X[mask] = self.handle_missing
+                encodings = _fit_encoding_fast(
+                    X_train, y_train, n_categories, self.smooth, y_mean
+                )
+            self._transform_X_ordinal(
+                X_out, X_ordinal, X_unknown_mask, test_idx, encodings, y_mean
+            )
+        return X_out
 
-        X_temp = check_array(X, dtype=None)
-        X = X_temp
+    def transform(self, X):
+        """Transform X with the target encoding.
 
-        n_samples, n_features = X.shape
-        X_int = np.zeros_like(X, dtype=int)
-        X_mask = np.ones_like(X, dtype=bool)
+        .. note::
+            `fit(X, y).transform(X)` does not equal `fit_transform(X, y)` because a
+            cross fitting scheme is used in `fit_transform` for encoding. See the
+            :ref:`User Guide <target_encoder>`. for details.
 
-        for i in range(n_features):
-            Xi = X[:, i]
-            valid_mask = np.in1d(Xi, self.categories_[i])
+        Parameters
+        ----------
+        X : array-like of shape (n_samples, n_features)
+            The data to determine the categories of each feature.
 
-            if not np.all(valid_mask):
-                if self.handle_unknown == "error":
-                    diff = np.unique(X[~valid_mask, i])
-                    raise ValueError(
-                        f"Found unknown categories {diff} in column {i} "
-                        "during transform."
-                    )
-                else:
-                    # Set the problematic rows to an acceptable value and
-                    # continue `The rows are marked `X_mask` and will be
-                    # removed later.
-                    X_mask[:, i] = valid_mask
-                    Xi = Xi.copy()
-                    Xi[~valid_mask] = self.categories_[i][0]
-            X_int[:, i] = self._label_encoders_[i].transform(Xi)
+        Returns
+        -------
+        X_trans : ndarray of shape (n_samples, n_features)
+            Transformed input.
+        """
+        X_ordinal, X_valid = self._transform(
+            X, handle_unknown="ignore", force_all_finite="allow-nan"
+        )
+        X_out = np.empty_like(X_ordinal, dtype=np.float64)
+        self._transform_X_ordinal(
+            X_out,
+            X_ordinal,
+            ~X_valid,
+            slice(None),
+            self.encodings_,
+            self.target_mean_,
+        )
+        return X_out
 
-        out = []
+    def _fit_encodings_all(self, X, y):
+        """Fit a target encoding with all the data."""
 
-        for j, cats in enumerate(self.categories_):
-            unqX = np.unique(X[:, j])
-            encoder = {x: 0 for x in unqX}
-            if self.clf_type in ["binary-clf", "regression"]:
-                for x in unqX:
-                    if x not in cats:
-                        Eyx = 0
-                    else:
-                        Eyx = self.Eyx_[j][x]
-                    lambda_n = lambda_(self.counter_[j][x], self.n_ / self.k_[j])
-                    encoder[x] = lambda_n * Eyx + (1 - lambda_n) * self.Ey_
-                x_out = np.zeros((len(X[:, j]), 1))
-                for i, x in enumerate(X[:, j]):
-                    x_out[i, 0] = encoder[x]
-                out.append(x_out.reshape(-1, 1))
-            if self.clf_type == "multiclass-clf":
-                x_out = np.zeros((len(X[:, j]), len(self.classes_)))
-                lambda_n = {x: 0 for x in unqX}
-                for x in unqX:
-                    lambda_n[x] = lambda_(self.counter_[j][x], self.n_ / self.k_[j])
-                for k, c in enumerate(np.unique(self.classes_)):
-                    for x in unqX:
-                        if x not in cats:
-                            Eyx = 0
-                        else:
-                            Eyx = self.Eyx_[c][j][x]
-                        encoder[x] = lambda_n[x] * Eyx + (1 - lambda_n[x]) * self.Ey_[c]
-                    for i, x in enumerate(X[:, j]):
-                        x_out[i, k] = encoder[x]
-                out.append(x_out)
-        out = np.hstack(out)
-        return out
+        check_consistent_length(X, y)
+        self._fit(X, handle_unknown="ignore", force_all_finite="allow-nan")
+
+        if self.target_type == "auto":
+            accepted_target_types = ("binary", "continuous")
+            inferred_type_of_target = type_of_target(y, input_name="y")
+            if inferred_type_of_target not in accepted_target_types:
+                raise ValueError(
+                    f"Target type was inferred to be {inferred_type_of_target!r}. Only"
+                    f" {accepted_target_types} are supported."
+                )
+            self.target_type_ = inferred_type_of_target
+        else:
+            self.target_type_ = self.target_type
+
+        if self.target_type_ == "binary":
+            y = LabelEncoder().fit_transform(y)
+        else:  # continuous
+            y = _check_y(y, y_numeric=True, estimator=self)
+
+        self.target_mean_ = np.mean(y)
+
+        X_ordinal, X_known_mask = self._transform(
+            X, handle_unknown="ignore", force_all_finite="allow-nan"
+        )
+        n_categories = np.fromiter(
+            (len(category_for_feature) for category_for_feature in self.categories_),
+            dtype=np.int64,
+            count=len(self.categories_),
+        )
+        if self.smooth == "auto":
+            y_variance = np.var(y)
+            print(X_ordinal)
+            self.encodings_ = _fit_encoding_fast_auto_smooth(
+                X_ordinal, y, n_categories, self.target_mean_, y_variance
+            )
+        else:
+            self.encodings_ = _fit_encoding_fast(
+                X_ordinal, y, n_categories, self.smooth, self.target_mean_
+            )
+
+        return X_ordinal, X_known_mask, y, n_categories
+
+    @staticmethod
+    def _transform_X_ordinal(
+        X_out, X_ordinal, X_unknown_mask, indices, encodings, y_mean
+    ):
+        """Transform X_ordinal using encodings."""
+        for f_idx, encoding in enumerate(encodings):
+            X_out[indices, f_idx] = encoding[X_ordinal[indices, f_idx]]
+            X_out[X_unknown_mask[:, f_idx], f_idx] = y_mean
+
+    def _more_tags(self):
+        return {"requires_y": True}